+++ release_notes.md
@@ -27,3 +27,5 @@
    - Azure.Identity (1.10.0 to 1.10.3)
    - Azure.Core (1.34.0 to 1.35.0)
 - Updating HostWarmupMiddleware to enable /warmup API call when testing locally using DebugPlaceholder or ReleasePlaceholder configuration.
+- Bug fix: Comparisons in the Azure Key Vault Secrets Repository are now case insensitive (#9644)
+  - This fixes a bug where keys could be automatically recreated if they had been manually added to Key Vault with all lowercase secret names.